+++ Application/src/main/java/ingredient/model/Ingredient.java
@@ -6,15 +6,19 @@ public class Ingredient {
     int ingredientId;
     String ingredientName;
     NutritionalFacts nutrition;
-    BigDecimal quantity;
     MeasurementUnit measurementUnit;
+
+
+    public Ingredient(String ingredientName, int ingredientId) {
+        this.ingredientName = ingredientName;
+        this.ingredientId = ingredientId;
+    }
 
     public Ingredient(int ingredientId, String ingredientName, NutritionalFacts nutrition, MeasurementUnit measurementUnit, BigDecimal quantity) {
         this.ingredientId = ingredientId;
         this.ingredientName = ingredientName;
         this.nutrition = nutrition;
         this.measurementUnit = measurementUnit;
-        this.quantity = quantity;
     }
 
     public String getIngredientName() {
@@ -41,14 +45,6 @@ public class Ingredient {
         this.nutrition = nutrition;
     }
 
-    public BigDecimal getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(BigDecimal quantity) {
-        this.quantity = quantity;
-    }
-
     public MeasurementUnit getMeasurementUnit() {
         return measurementUnit;
     }
@@ -57,4 +53,13 @@ public class Ingredient {
         this.measurementUnit = measurementUnit;
     }
 
+    @Override
+    public String toString() {
+        return "Ingredient{" +
+                "ingredientId=" + ingredientId +
+                ", ingredientName='" + ingredientName + '\'' +
+                ", nutrition=" + nutrition +
+                ", measurementUnit=" + measurementUnit +
+                '}';
+    }
 }