+++ Application/src/main/java/ingredient/model/IngredientsInventory.java
@@ -1,36 +1,51 @@
 package ingredient.model;
 
+import recipe.model.Recipe;
+
 import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public class IngredientsInventory {
-    List<Ingredient> inventory;
+    private Map<Integer, Ingredient> inventory;
+    int ingredientID;
 
-    public IngredientsInventory(List<Ingredient> inventory) {
+    public IngredientsInventory(){
+
+        inventory = new HashMap<>();
+        ingredientID = 1;
+    }
+    public IngredientsInventory(Map<Integer, Ingredient> inventory) {
         this.inventory = inventory;
     }
 
-    public List<Ingredient> getInventory() {
+    public Map<Integer, Ingredient> getInventory() {
         return inventory;
     }
 
-    public void setInventory(List<Ingredient> inventory) {
+    public void setInventory(Map<Integer, Ingredient> inventory) {
         this.inventory = inventory;
     }
 
-    public void addIngredient(Ingredient ingredient) {
-        this.inventory.add(ingredient);
+    public Ingredient addIngredient(String ingredientName) {
+        Ingredient x = new Ingredient(ingredientName,ingredientID);
+        inventory.put(ingredientID, x);
+        ingredientID++;
+        return x;
     }
 
     public void deleteIngredient(Ingredient ingredient) {
-        this.inventory.remove(ingredient);
+        this.inventory.remove(ingredient.getIngredientId());
     }
 
     public void updateIngredient(Ingredient ingredient) {
-        this.inventory.set(inventory.indexOf(ingredient), ingredient);
+        //this.inventory.set(inventory.indexOf(ingredient), ingredient);
     }
 
-    public void getIngredientById(int ingredientId) {
+    public Ingredient getIngredientById(int ingredientId) {
+        return inventory.get(ingredientId);
         //Implement later
     }
 
@@ -40,5 +55,19 @@ public class IngredientsInventory {
 
     public void createIngredient(String name, NutritionalFacts nutritionalFacts, BigDecimal quantity, MeasurementUnit measurementUnit) {
         //Implement later
+    }
+
+    public void createIngredient(String name) {
+        Ingredient ingredient = new Ingredient(name,ingredientID);
+        ingredientID++;
+
+    }
+
+    public List<Ingredient> getIngredients(List<Integer> ingredientList) {
+        List<Ingredient> ingredients = new ArrayList<>();
+        for(Integer ingredientID: ingredientList){
+            ingredients.add(getIngredientById(ingredientID));
+        }
+        return ingredients;
     }
 }